app/tools/requirements-tools.txt
networkx>=3.2
pyvis>=0.3.2
httpx>=0.27
python-dotenv>=1.0